+++ jobseekr/api/models/meetupModel.js
@@ -2,30 +2,30 @@ const mongoose = require('mongoose');
 const Users = require('./userModels');
 
 const MeetupSchema = mongoose.Schema(
-    {
-        dateOfEvent: {
-            type: Date,
-            required: true,
-        },
-        eventName: {
-            type: String,
-            required: true,
-            minlength: 1,
-            maxlength: 40,
-        },
-        linkToEvent: {
-            type: String
-        },
-        notes: {
-            type: String,
-            maxlength: 100,
-        },
-        user: {
-          type: ObjectId,
-          ref: 'Users',
-        },
+  {
+    dateOfEvent: {
+      type: Date,
+      required: true,
     },
-    { timestamps: true }
+    eventName: {
+      type: String,
+      required: true,
+      minlength: 1,
+      maxlength: 40,
+    },
+    linkToEvent: {
+      type: String,
+    },
+    notes: {
+      type: String,
+      maxlength: 100,
+    },
+    user: {
+      type: ObjectId,
+      ref: 'Users',
+    },
+  },
+  { timestamps: true },
 );
 // add validation for link string types
