+++ jobseekr/src/components/AllComponents.jsx
@@ -1,0 +1,4 @@
+export { default as Header } from './Header';
+export { default as AddJob } from './AddJob';
+export { default as AddList } from './AddList';
+export { default as Carousel } from './Carousel';
