--- jobseekr/src/less/settings.css
@@ -1,14 +1,0 @@
-.navbar {
-  border-radius: 0px;
-  margin-bottom: -2%;
-}
-.form-group {
-  padding-left: 10px;
-  padding-right: 10px;
-}
-.btn-default {
-  margin: 10px;
-}
-.settingsWrapper {
-  height: 1600px;
-}
