+++ jobseekr/src/routes/AllRoutes.jsx
@@ -1,0 +1,10 @@
+export { default as Billing } from './Billing';
+export { default as Contributions } from './Contributions';
+export { default as Jobs } from './Jobs';
+export { default as LandingPage } from './LandingPage';
+export { default as Login } from './Login';
+export { default as SignOut } from './SignOut';
+export { default as SignUp } from './SignUp';
+export { default as MeetUps } from './MeetUps';
+export { default as Settings } from './Settings';
+
