+++ jobseekr/src/routes/Billing.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Header } from '../components/AllComponents';
+
+const Billing = () => (
+  <div className="parent">
+    <Header />
+    <div className="BillingWrapper">
+      <div>
+        <h4>hello im billing</h4>
+      </div>
+    </div>
+  </div>
+);
+
+export default Billing;
