+++ jobseekr/src/routes/Contributions.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Header } from '../components/AllComponents';
+
+const Contributions = () => (
+  <div className="parent">
+    <Header />
+    <div className="ContributionsWrapper">
+      <div>
+        <h4>hello im Contributions</h4>
+      </div>
+    </div>
+  </div>
+);
+
+export default Contributions;
