+++ jobseekr/src/routes/MeetUps.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Header } from '../components/AllComponents';
+
+const Contributions = () => (
+  <div className="parent">
+    <Header />
+    <div className="MeetupsWrapper">
+      <div>
+        <h4>hello im Meetups</h4>
+      </div>
+    </div>
+  </div>
+);
+
+export default Contributions;
