+++ jobseekr/src/routes/SignOut.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Header } from '../components/AllComponents';
+
+const SignOut = () => (
+  <div className="parent">
+    <Header />
+    <div className="SignOutWrapper">
+      <div>
+        <h4>hello im Contributions</h4>
+      </div>
+    </div>
+  </div>
+);
+
+export default SignOut;
