--- jobseekr/src/routes/billing/index.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import Header from '../../components/header';
-
-const Billing = () =>
-  (
-    <div className="parent">
-      <Header />
-      <div className="BillingWrapper">
-        <div>
-          <h4>hello im billing
-          </h4>
-        </div>
-      </div>
-    </div>
-  );
-
-export default Billing;
