--- jobseekr/src/routes/contributions/index.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import Header from '../../components/header';
-
-const Contributions = () =>
-  (
-    <div className="parent">
-      <Header />
-      <div className="ContributionsWrapper">
-        <div>
-          <h4>hello im Contributions
-          </h4>
-        </div>
-      </div>
-    </div>
-  );
-
-export default Contributions;
