--- jobseekr/src/routes/meetUps/index.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import Header from '../../components/header';
-
-const Contributions = () =>
-  (
-    <div className="parent">
-      <Header />
-      <div className="MeetupsWrapper">
-        <div>
-          <h4>hello im Meetups
-          </h4>
-        </div>
-      </div>
-    </div>
-  );
-
-export default Contributions;
-
