--- jobseekr/src/routes/signOut/index.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import Header from '../../components/header';
-
-const SignOut = () =>
-  (
-    <div className="parent">
-      <Header />
-      <div className="SignOutWrapper">
-        <div>
-          <h4>hello im Contributions
-          </h4>
-        </div>
-      </div>
-    </div>
-  );
-
-export default SignOut;
